refactor(@lib/create-event-handler): rename file
