+++ lib/create-event-handlers/src/create-event-handler/create-event-handler.js
@@ -1,7 +1,7 @@
 import { createValidate } from '@lib/create-validate';
 
 import { createValidateHandlerResponse } from './create-validate-handler-response';
-import { createToHandlerEvent } from './createto-handler-event';
+import { createToHandlerEvent } from './create-to-handler-event';
 import { createHandleResponse } from './create-handle-response';
 
 export const createEventHandler = ({
